It is almost like working

+++ execution-engine/src/models/execution_state.py
@@ -10,7 +10,7 @@ class ExecutionState(BaseModel):
     bot_id: int # id of the bot for execution
     execution_id: int # id of the execution
     executing_node_id: str # id of the node being executed
-    variable_values: dict[str, str|int]
+    variable_values: dict[str, str|float]
 
 
 class RunTimeExecutionState(ExecutionState):
